+++ lib/utils/ProductHelper.dart
@@ -1,15 +1,51 @@
 import '../model/Product.dart';
-import '../model/Ingredient.dart';
+import '../model/ProductImage.dart';
 import '../model/User.dart';
+import '../model/Ingredient.dart';
 
-class IngredientHelper {
+
+class ProductHelper {
+
+  /// reduce the set of images of the product depending on the given language.
+  static void removeImages(Product product, String language) {
+    for (String field in ProductImage.FIELDS) {
+      if (product.selectedImages.list.any(
+              (i) => i.field == field && i.language == language)) {
+        product.selectedImages.list.removeWhere(
+                (i) => i.field == field && i.language != language);
+      }
+    }
+  }
+
+  /// prepare the product name for the given product.
+  /// remove name strings of other languages.
+  static void prepareProductName(Product product, String lang) {
+    // choose the language specific product name
+    switch (lang) {
+      case User.LANGUAGE_DE:
+        product.productName = product.productNameDE ?? product.productName;
+        break;
+      case User.LANGUAGE_EN:
+        product.productName = product.productNameEN ?? product.productName;
+        break;
+      case User.LANGUAGE_FR:
+        product.productName = product.productNameFR ?? product.productName;
+        break;
+      default:
+        break;
+    }
+
+    // remove unused fields
+    product.productNameDE = null;
+    product.productNameEN = null;
+    product.productNameFR = null;
+  }
 
   /// prepare the ingredients for the given product.
   /// parse the language specific ingredient text into ingredient objects.
   static void parseIngredients(Product product, String lang) {
-
     // choose the language specific string of ingredients
-    switch(lang) {
+    switch (lang) {
       case User.LANGUAGE_DE:
         product.ingredientsText = product.ingredientsTextDE;
         break;
@@ -30,7 +66,6 @@ class IngredientHelper {
     product.ingredientsTextFR = null;
 
     if (product.ingredientsText != null && product.ingredientsText.isNotEmpty) {
-
       // find the names of all ingredients within the given string.
       // An item can contain 0-n spaces or hyphens "-"
       // An element may contain 1-n letters
@@ -43,8 +78,7 @@ class IngredientHelper {
         String name = m.group(0).trim();
 
         // avoid empty ingredients
-        if (name.isNotEmpty ) {
-
+        if (name.isNotEmpty) {
           // remove numbers with percent (e.g. 12%)
           name = name.replaceAll(new RegExp(r"(([0-9])*%)"), "").trim();